+++ src/app/frame_context.rs
@@ -11,7 +11,7 @@ use crate::ecs::resource::gizmo::{
 use crate::ecs::resource::DebugViewState;
 use crate::ecs::resource::{Camera, GridMeshData, LightState};
 use crate::ecs::world::{ResMut, ResRef, World};
-use crate::render::RenderBackend;
+use crate::render::{BillboardBackend, RenderBackend};
 use crate::vulkanr::command::RRCommandPool;
 use crate::vulkanr::device::RRDevice;
 use crate::vulkanr::renderer::onion_skin_buffers::OnionSkinGpuState;
