+++ src/render/mod.rs
@@ -1,6 +1,6 @@
 pub mod backend;
 
-pub use backend::{MeshId, RenderBackend};
+pub use backend::{BillboardBackend, MeshId, RenderBackend};
 pub use thyllore_render_core::{
     BufferHandle, BufferMemoryType, FrameUBO, IndexBufferHandle, MaterialUBO, ObjectUBO,
     VertexBufferHandle,
